# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Coinroll MEME Launchpad API",
    "description": "MEME代币发射平台后端API接口文档，用于复刻后端服务",
    "version": "1.0.0",
    "contact": {
      "name": "Coinroll Team"
    }
  },
  "servers": [
    {
      "url": "https://api-dev.coinroll.io/api/v1",
      "description": "开发环境"
    },
    {
      "url": "https://api.coinroll.io/api/v1",
      "description": "生产环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token认证"
      }
    },
    "schemas": {
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "pageSize": {"type": "integer"},
          "pageNo": {"type": "integer"},
          "total": {"type": "integer"},
          "totalPage": {"type": "integer"}
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "address": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"},
          "username": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "avatar": {"type": "string", "format": "uri"},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
...
      "OverviewStats": {
        "type": "object",
        "properties": {
          "createdTokens": {"type": "integer"},
          "heatedTokens": {"type": "integer"},
          "ownedTokens": {"type": "integer"},
          "pendingUnlockTokens": {"type": "integer"},
          "totalTradeBnb": {"type": "string"}
        }
      },
...
    }
  },
  "paths": {
    "/user/sign-msg": {
      "get": {
        "tags": ["Auth"],
        "summary": "获取签名消息",
        "description": "获取用于钱包签名登录的消息",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"},
            "description": "钱包地址"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {"type": "string", "description": "需要签名的消息"},
                        "nonce": {"type": "string"},
                        "expiry": {"type": "integer"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/wallet-login": {
      "post": {
        "tags": ["Auth"],
        "summary": "钱包登录",
        "description": "使用钱包签名进行登录认证",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["address", "signature"],
                "properties": {
                  "address": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"},
                  "signature": {"type": "string", "description": "钱包签名"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {"type": "string", "description": "访问令牌"},
                        "refreshToken": {"type": "string"},
                        "expiresIn": {"type": "integer"},
                        "tokenType": {"type": "string", "default": "Bearer"},
                        "user": {"$ref": "#/components/schemas/User"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/refresh-token": {
      "post": {
        "tags": ["Auth"],
        "summary": "刷新令牌",
        "description": "使用刷新令牌获取新的访问令牌",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {"type": "string"},
                        "refreshToken": {"type": "string"},
                        "expiresIn": {"type": "integer"},
                        "tokenType": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/overview-stats": {
      "get": {
        "tags": ["Profile"],
        "summary": "获取用户概览统计",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "schema": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {"$ref": "#/components/schemas/OverviewStats"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/following-list": {
      "get": {
        "tags": ["Profile"],
        "summary": "获取关注列表",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/user/followers-list": {
      "get": {
        "tags": ["Profile"],
        "summary": "获取粉丝列表",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "search", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/token/token-list": {
      "get": {
        "tags": ["Token"],
        "summary": "获取代币列表",
        "parameters": [
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "页码"},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 10}, "description": "每页数量"},
          {"name": "sort", "in": "query", "schema": {"type": "string"}, "description": "排序字段"},
          {"name": "lunchMode", "in": "query", "schema": {"type": "integer"}, "description": "发射模式"},
          {"name": "status", "in": "query", "schema": {"type": "string"}, "description": "状态"},
          {"name": "launchTimeStart", "in": "query", "schema": {"type": "integer"}, "description": "发射时间开始"},
          {"name": "launchTimeEnd", "in": "query", "schema": {"type": "integer"}, "description": "发射时间结束"}
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {
                      "allOf": [
                        {"$ref": "#/components/schemas/PaginatedResponse"},
                        {
                          "type": "object",
                          "properties": {
                            "result": {
                              "type": "array",
                              "items": {"$ref": "#/components/schemas/TokenListItem"}
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/token/detail": {
      "get": {
        "tags": ["Token"],
        "summary": "获取代币详情",
        "parameters": [
          {
            "name": "tokenAddress",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"},
            "description": "代币合约地址"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {"$ref": "#/components/schemas/TokenDetail"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/token/trending-token": {
      "get": {
        "tags": ["Token"],
        "summary": "获取趋势代币",
        "parameters": [
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "lunchMode", "in": "query", "schema": {"type": "integer"}},
          {"name": "launchTimeStart", "in": "query", "schema": {"type": "integer"}},
          {"name": "launchTimeEnd", "in": "query", "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/token/holders": {
      "get": {
        "tags": ["Token"],
        "summary": "获取代币持有者列表",
        "parameters": [
          {
            "name": "tokenAddress",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "pattern": "^0x[a-fA-F0-9]{40}$"}
          },
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {"type": "integer"},
                    "message": {"type": "string"},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/TokenHolder"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/token/create-token": {
      "post": {
        "tags": ["Token"],
        "summary": "创建代币",
        "description": "创建新代币，返回合约调用所需的参数和签名",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateTokenInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CreateTokenResponse"}
              }
            }
          }
        }
      }
    },
...
    "/token/favorite": {
      "post": {
        "tags": ["Token"],
        "summary": "收藏代币",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tokenId"],
                "properties": {
                  "tokenId": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "成功"}
        }
      }
    },
    "/token/unfavorite": {
      "post": {
        "tags": ["Token"],
        "summary": "取消收藏代币",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tokenId"],
                "properties": {
                  "tokenId": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "成功"}
        }
      }
    },
...
    "/trade/upcoming-token": {
      "get": {
        "tags": ["Trade"],
        "summary": "获取即将上线的代币",
        "parameters": [
          {"name": "pn", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "ps", "in": "query", "schema": {"type": "integer", "default": 30}}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      }
    },
...
  }
}